refactor(gateway): move Config-level tests to config/tests.rs (GW-2 S9)
Use the Rust 'fichier frere' pattern: #[cfg(test)] mod tests; in
config.rs activates the sibling config/tests.rs only in test builds.
